Init App2UseMemo
