+++ src/components2/App2UseMemo.js
@@ -1,0 +1,15 @@
+import React, {Component} from 'react'
+
+class App2UseMemo extends Component {
+
+  render() {
+    return (
+
+      <div className="container">
+
+      </div>
+    );  
+  }
+}
+
+export default App2UseMemo;
